Allow to override dynamic tooltip styles
Since tooltip bubble width, heigh and position are dynamically
calculated, it doesn't allow to make it statically sized, due to this
it was not possible to change the bubble to take has much width as
possible.
This became clear when the highlighted element was in the middle of the
screen making the bubble be confined to less than half of the screen, so
with some considerable amount of text, it would look rather weird.

Signed-off-by: Marcos Loureiro <mloureiro.dev@gmail.com>